add: Carousel style.ts 분리
